+++ Packages/com.chark.game-management/Runtime/Entities/EntityManager.cs
@@ -1,19 +1,20 @@
 ﻿using System;
 using System.Collections.Generic;
-using UnityEngine;
+using CHARK.GameManagement.Settings;
+using CHARK.GameManagement.Utilities;
 
 namespace CHARK.GameManagement.Entities
 {
     internal sealed class EntityManager : IEntityManager
     {
-        private readonly List<object> entities = new List<object>();
-        private readonly bool isVerboseLogging;
+        private readonly List<object> entities = new();
+        private readonly IGameManagerSettingsProfile profile;
 
         public IReadOnlyList<object> Entities => entities;
 
-        public EntityManager(bool isVerboseLogging = false)
+        public EntityManager(IGameManagerSettingsProfile profile)
         {
-            this.isVerboseLogging = isVerboseLogging;
+            this.profile = profile;
         }
 
         public bool AddEntity<TEntity>(TEntity entity) where TEntity : class
@@ -21,17 +22,17 @@ namespace CHARK.GameManagement.Entities
 #if UNITY_EDITOR
             if (entity == null)
             {
-                Debug.LogError("Entity cannot be null");
+                Logging.LogError("Entity cannot be null", GetType());
                 return false;
             }
 #endif
 
-            if (isVerboseLogging)
+            if (profile.IsVerboseLogging)
             {
                 var entityType = entity.GetType();
                 var entityName = entityType.Name;
 
-                Debug.Log($"Adding entity {entityName}");
+                Logging.LogDebug($"Adding entity {entityName}", GetType());
             }
 
             entities.Add(entity);
@@ -43,7 +44,7 @@ namespace CHARK.GameManagement.Entities
 #if UNITY_EDITOR
             if (entity == null)
             {
-                Debug.LogError("Entity cannot be null");
+                Logging.LogError("Entity cannot be null", GetType());
                 return false;
             }
 #endif
@@ -57,12 +58,12 @@ namespace CHARK.GameManagement.Entities
                     continue;
                 }
 
-                if (isVerboseLogging)
+                if (profile.IsVerboseLogging)
                 {
                     var entityType = entity.GetType();
                     var entityName = entityType.Name;
 
-                    Debug.Log($"Removing entity {entityName}");
+                    Logging.LogDebug($"Removing entity {entityName}", GetType());
                 }
 
                 entities.RemoveAt(index);