+++ Packages/com.chark.game-management/Runtime/GameManager.cs
@@ -1,8 +1,11 @@
-﻿using CHARK.GameManagement.Assets;
+﻿using System;
+using CHARK.GameManagement.Assets;
 using CHARK.GameManagement.Entities;
 using CHARK.GameManagement.Messaging;
+using CHARK.GameManagement.Settings;
 using CHARK.GameManagement.Storage;
 using CHARK.GameManagement.Systems;
+using CHARK.GameManagement.Utilities;
 using UnityEngine;
 
 namespace CHARK.GameManagement
@@ -19,6 +22,8 @@ namespace CHARK.GameManagement
             DefaultGameStorage.Instance;
 #endif
 
+        private static GameManagerSettings Settings => GameManagerSettings.Instance;
+
         private bool isSystemsInitialized;
 
         private IGameStorage runtimeStorage;
@@ -26,69 +31,12 @@ namespace CHARK.GameManagement
         private IEntityManager entityManager;
         private IMessageBus messageBus;
 
-        [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.SubsystemRegistration)]
-        private static void OnSubsystemRegistration()
-        {
-            OnInstantiate(RuntimeInitializeLoadType.SubsystemRegistration);
-        }
-
-        [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.AfterAssembliesLoaded)]
-        private static void OnAfterAssembliesLoaded()
-        {
-            OnInstantiate(RuntimeInitializeLoadType.AfterAssembliesLoaded);
-        }
-
-        [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.AfterSceneLoad)]
-        private static void OnAfterSceneLoad()
-        {
-            OnInstantiate(RuntimeInitializeLoadType.AfterSceneLoad);
-        }
-
-        [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.BeforeSceneLoad)]
-        private static void OnBeforeSceneLoad()
-        {
-            OnInstantiate(RuntimeInitializeLoadType.BeforeSceneLoad);
-        }
-
-        [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.BeforeSplashScreen)]
-        private static void OnBeforeSplashScreen()
-        {
-            OnInstantiate(RuntimeInitializeLoadType.BeforeSplashScreen);
-        }
-
-        private static void OnInstantiate(RuntimeInitializeLoadType targetLoadType)
-        {
-            var settings = GameManagerSettings.Instance;
-            if (settings.IsInstantiateAutomatically == false)
-            {
-                return;
-            }
-
-            var settingsLoadType = settings.LoadType;
-            if (settingsLoadType != targetLoadType)
-            {
-                return;
-            }
-
-            if (settings.TryGetGameManagerPrefab(out var prefab) == false)
-            {
-                Debug.LogError(
-                    $"Game Manager Prefab is not set in {nameof(GameManagerSettings)} settings"
-                );
-
-                return;
-            }
-
-            var newGameManager = Instantiate(prefab);
-            newGameManager.name = newGameManager.GetGameManagerName();
-        }
-
         private void Awake()
         {
             if (currentGameManager && currentGameManager != this)
             {
                 var gameManagerTypeName = GetType().Name;
-                Debug.LogWarning($"{gameManagerTypeName} is already initialized", this);
+                Logging.LogWarning($"{gameManagerTypeName} is already initialized", this);
                 Destroy(gameObject);
                 return;
             }
@@ -101,8 +49,13 @@ namespace CHARK.GameManagement
             {
                 InitializeGameManager();
 
-                var settings = GetGameManagerSettings();
-                if (settings.IsDontDestroyOnLoad)
+                var profile = Settings.ActiveProfile;
+                if (profile.IsVerboseLogging)
+                {
+                    Logging.LogDebug($"{GetGameManagerName()} initialized", this);
+                }
+
+                if (profile.IsDontDestroyOnLoad)
                 {
                     DontDestroyOnLoad(gameObject);
                 }
@@ -123,10 +76,19 @@ namespace CHARK.GameManagement
         {
             OnBeforeDestroy();
 
-            var systems = entityManager.GetEntities<ISystem>();
-            foreach (var system in systems)
+            if (entityManager != null)
             {
-                RemoveSystem(system);
+                var systems = entityManager.GetEntities<ISystem>();
+                foreach (var system in systems)
+                {
+                    RemoveSystem(system);
+                }
+            }
+
+            var profile = Settings.ActiveProfile;
+            if (profile.IsVerboseLogging)
+            {
+                Logging.LogDebug($"{GetGameManagerName()} disposed", this);
             }
         }
 
@@ -152,20 +114,11 @@ namespace CHARK.GameManagement
         }
 
         /// <returns>
-        /// Custom <see cref="IGameManagerSettings"/> implementation. Defaults to
-        /// <see cref="GameManagerSettings.Instance"/>.
-        /// </returns>
-        protected virtual IGameManagerSettings GetGameManagerSettings()
-        {
-            return GameManagerSettings.Instance;
-        }
-
-        /// <returns>
         /// Name of this game manager.
         /// </returns>
         protected virtual string GetGameManagerName()
         {
-            return name;
+            return GetType().Name;
         }
 
         /// <summary>
@@ -212,8 +165,8 @@ namespace CHARK.GameManagement
         /// </returns>
         protected virtual IEntityManager CreateEntityManager()
         {
-            var settings = GetGameManagerSettings();
-            return new EntityManager(settings.IsVerboseLogging);
+            var profile = Settings.ActiveProfile;
+            return new EntityManager(profile);
         }
 
         /// <returns>
@@ -235,6 +188,7 @@ namespace CHARK.GameManagement
 
         private void InitializeCore()
         {
+            name = GetGameManagerName();
             runtimeStorage = CreateRuntimeGameStorage();
             resourceLoader = CreateResourceLoader();
             entityManager = CreateEntityManager();
@@ -254,13 +208,13 @@ namespace CHARK.GameManagement
                     continue;
                 }
 
-                var settings = GetGameManagerSettings();
-                if (settings.IsVerboseLogging)
+                var profile = Settings.ActiveProfile;
+                if (profile.IsVerboseLogging)
                 {
                     var systemType = entity.GetType();
                     var systemName = systemType.Name;
 
-                    Debug.Log($"Initializing system {systemName}", this);
+                    Logging.LogDebug($"Initializing system {systemName}", this);
                 }
 
                 system.OnInitialized();
@@ -280,8 +234,14 @@ namespace CHARK.GameManagement
 
             if (currentGameManager == false)
             {
-                Debug.LogError($"{nameof(GameManager)} is not initialized");
-                Debug.Break();
+                throw new Exception(
+                    $"{nameof(GameManager)} is not initialized."
+                    + $" Did you forget to add it to one of your scenes?"
+                    + $" If you're using automatic instantiation "
+                    + $" ({nameof(GameManagerSettingsProfile.IsInstantiateAutomatically)}),"
+                    + $" make sure the {nameof(GameManagerSettings)} contains a valid and active"
+                    + $" profile with a set {nameof(GameManager)} prefab."
+                );
             }
 #endif